requirements.txt
Flask==2.1.2
tensorflow==2.10.0
Pillow==9.2.0
numpy==1.21.0
